feat: add base socket and tcp socket class
