+++ include/musen/socket/tcp_socket.hpp
@@ -18,17 +18,20 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef MUSEN__MUSEN_HPP_
-#define MUSEN__MUSEN_HPP_
+#ifndef MUSEN__TCP_SOCKET_HPP_
+#define MUSEN__TCP_SOCKET_HPP_
 
-#include "./broadcaster/base_broadcaster.hpp"
-#include "./broadcaster/broadcaster.hpp"
-#include "./broadcaster/string_broadcaster.hpp"
-#include "./listener/base_listener.hpp"
-#include "./listener/listener.hpp"
-#include "./listener/string_listener.hpp"
-#include "./socket/base_socket.hpp"
-#include "./socket/tcp_socket.hpp"
-#include "./socket/udp_socket.hpp"
+#include "./base_socket.hpp"
 
-#endif  // MUSEN__MUSEN_HPP_
+namespace musen
+{
+
+class TcpSocket : public BaseSocket
+{
+public:
+  bool connect();
+};
+
+}  // namespace musen
+
+#endif  // MUSEN__TCP_SOCKET_HPP_
